lttng: preserve old modular only restriction
In commit 240311e62d185d1b5bf24c385c46bb9cdda9f122
("lttng: delete hook to staging/lttng2; we re-use staging/lttng")
we deleted lttng2-added-lttng-Kconfig-for-in-kernel-building.patch
which forced modular only building of lttng [in order to be aligned
with the out of tree build the ltt tools expected].  Restore the
part of that patch which added this restriction, along with the
assoc. chunk of the old commit log and authorship.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>