emgd: Fixes for 3.10 kernel
1/2 [
Author: Tom Zanussi
Email: tom.zanussi@linux.intel.com
Subject: emgd: Fixes for 3.10 kernel
Date: Sun, 15 Sep 2013 16:39:51 -0500

Various fixes to get emgd-1.18 compiling with the 3.10 kernel. The
main changes were to replace the defunct create_proc_read_entry() etc
with proc_create_data() and PDE_DATA() and fix up proc entry removal
in light of the fact that proc_dir_entry is now opaque and used to
contain the name.

The proc files we need to worry about are handled by the above changes
- there are a couple sections of proc removal and write_proc() code
that were removed because they're only used by DEBUG code and we don't
care about fixing up debug code.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
]

2/2 [
Author: Tom Zanussi
Email: tom.zanussi@linux.intel.com
Subject: emgd-1.18: Fix ioctl handling
Date: Mon, 7 Oct 2013 07:52:29 -0500

commit e4fda9f [drm: Perform ioctl command validation on the stored
kernel values] causes a couple problems that need to be fixed.

One is that ioctl.cmd_drv needs to be given an appropriate value
instead of 0 since it's now used as the cmd value in the handler -
EMGD_IOCTL_DEF_DRV() is introduced and used instead of
EMGD_IOCTL_DEF().

The other problem is that the above exposes the fact that the cmds
that userspace is sending are not what the kernel says it's expecting
e.g. the kernel says IOR but what's being passed in is IOWR, which
causes different (correct) kernel handling, but that doesn't match the
kernel code.  So the kernel code is wrong, and if not corrected makes
video inoperable.  Since we don't have access to the userland code
that sends the commands, we need to change the affected kernel ioctl
values to match.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>