SMP Guest WRLinux 4.3 crashes when HV debug shell is exited
CQID: WIND00431175

1/1 [
Author: gcaron
Email: Guillaume.Caron@windriver.com
Subject: WIND00431175: SMP Guest WRLinux 4.3 crashes when HV debug shell is exited.
Date: Thu, 10 Oct 2013 10:06:14 -0400

In WRLinux 4.3 SMP guests, the BoardConfig area is remapped to the top of the
32-bit guest virtual address space and internal pointers in the BoardConfig area
are modified to account for this remap. Unfortunately, this pointer fixup is
applied every time the SMP WRLinux 4.3 guest is (re)started.
This patch resolves the issue.
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>