arch: delete empty omap.scc file
We intend to make inclusion of an arch a hard requirement of
each BSP, in order to avoid arch conditional statements in
scc files (which are problematic during tree generation where
no specific arch is really set).  So we'll keep the arch dir
in the commonly understood sense of being arm/mips/powerpc/x86
etc, in keeping with the kernel.

Note that there isn't any removal of "include arch/omap.scc"
here in this commit, since there isn't any to remove.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>