xen/blkback: Check device permissions before allowing OP_DISCARD -  linux - CVE-2013-2140
CQID: WIND00439190

1/1 [
Author: "yanjun.zhu"
Email: yanjun.zhu@windriver.com
Subject: xen/blkback: Check device permissions before allowing OP_DISCARD
Date: Mon, 21 Oct 2013 14:49:36 +0800

commit 604c499cbbcc3d5fe5fb8d53306aa0fae1990109 upstream.

We need to make sure that the device is not RO or that
the request is not past the number of sectors we want to
issue the DISCARD operation for.

This fixes CVE-2013-2140.

Cc: stable@vger.kernel.org
Acked-by: Jan Beulich <JBeulich@suse.com>
Acked-by: Ian Campbell <Ian.Campbell@citrix.com>
[v1: Made it pr_warn instead of pr_debug]
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>

Signed-off-by: yanjun.zhu <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>