ocf: remove the unneeded invocation of daemonize()
1/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: ocf: remove the unneeded invocation of daemonize()
Date: Mon, 21 Oct 2013 11:12:37 +0800

In the current kernel, all kthread are created by the kthreadd thread
and the invocation of function daemonize() for the kthread function is
not needed anymore. Actually function daemonize() has already been killed
by commit c4144670 (kill daemonize()) in the latest kernel.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: ocf/talitos: add FSL_SOC dependency
Date: Mon, 21 Oct 2013 11:12:38 +0800

The talitos crypto engine only exist on FSL PPC SoCs and its driver
is not buildable for other archs. So add the FSL_SOC dependency for it.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

3/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: ocf: use the correct kthread helper functions
Date: Mon, 21 Oct 2013 11:12:39 +0800

In the current kernel, the function kernel_thread() is not the right
API to create a kthread. So we should use the correct kthread helper
functions to create and stop a kthread.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

4/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: ocf: pass the correct size argument to memset
Date: Mon, 21 Oct 2013 11:12:40 +0800

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

5/5 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: ocf: include <linux/irqreturn.h> in ocf-compat.h
Date: Mon, 21 Oct 2013 11:12:41 +0800

In order to be compatible with the old kernel, the ocf-compat.h will
define the IRQ_NONE and IRQ_RETVAL to NULL if they are not defined
in other header files. So we should make sure that the file in which
these macros are defined must be included before this check. The file
"linux/irqreturn.h" was introduced only after v2.6.17, so also add
a kernel version check for this inclusion as what it did in other
part of this file.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>