fix ftrace_stress failed in bcm-xlp RT Kernel
CQID: WIND00440700

1/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: rt, powerpc: avoid to run the irq work in irq context
Date: Wed, 30 Oct 2013 13:36:50 +0800

For rt kernel, most of the irq handlers run in thread context, so
we also need to make sure that the irq work also run in thread context
in order to fix the call trace like the following. The patch
(x86-no-perf-irq-work-rt.patch) has made corresponding changes for
the x86 and common parts. This is just for the powerpc part.

  BUG: sleeping function called from invalid context at kernel/rtmutex.c:659
  in_atomic(): 1, irqs_disabled(): 1, pid: 625, name: ftrace_current_
  Preemption disabled at:[<  (null)>]   (null)

  CPU: 0 PID: 625 Comm: ftrace_current_ Not tainted 3.10.15-rt11+ #21
  Call Trace:
  [c7a9fcf0] [c0008d1c] show_stack+0xfc/0x1c0 (unreliable)
  [c7a9fd40] [c06ffe14] rt_spin_lock+0x34/0x90
  [c7a9fd50] [c0080498] __wake_up+0x38/0x70
  [c7a9fd70] [c0109268] __irq_work_run+0xc8/0x110
  [c7a9fd90] [c000ae58] timer_interrupt+0x238/0x330
  [c7a9fdc0] [c001188c] ret_from_except+0x0/0x18
  --- Exception: 901 at sub_preempt_count+0x14/0xe0
  LR = arch_irq_work_raise+0x50/0xa0
  [c7a9fe80] [ef830200] 0xef830200 (unreliable)
  [c7a9fe90] [c000abd0] arch_irq_work_raise+0x50/0xa0
  [c7a9fea0] [c01094b4] irq_work_queue+0x114/0x170
  [c7a9feb0] [c00ec6e0] ring_buffer_unlock_commit+0x150/0x190
  [c7a9fed0] [c00f3ff0] trace_buffer_unlock_commit+0x20/0x60
  [c7a9fef0] [c0002b84] ftrace_raw_event_sys_enter+0xe4/0xf0
  [c7a9ff20] [c00047b8] do_syscall_trace_enter+0x148/0x190
  [c7a9ff40] [c001136c] syscall_dotrace+0x5c/0xc8
  --- Exception: c00 at 0xff22570

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/2 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: rt: Revert "ring-buffer: Do not use schedule_work_on() for current CPU"
Date: Tue, 29 Oct 2013 19:35:29 +0800

The commit f5eb5588262c(ring-buffer: Do not use schedule_work_on() for
current CPU) introduces the following call trace for the rt kernel:

  BUG: sleeping function called from invalid context at kernel/rtmutex.c:659
  in_atomic(): 1, irqs_disabled(): 0, pid: 616, name: ftrace_buffer_s
  Preemption disabled at:[<  (null)>]   (null)

  CPU: 0 PID: 616 Comm: ftrace_buffer_s Not tainted 3.10.15-rt11+ #22
  Call Trace:
  [c751bd80] [c0008d1c] show_stack+0xfc/0x1c0 (unreliable)
  [c751bdd0] [c06ffdd4] rt_spin_lock+0x34/0x90
  [c751bde0] [c0115134] free_hot_cold_page+0xb4/0x2e0
  [c751be20] [c00ea760] free_buffer_page+0x20/0x40
  [c751be30] [c00ed680] rb_update_pages+0x230/0x310
  [c751be60] [c00edba8] ring_buffer_resize+0x408/0x5a0
  [c751bea0] [c00f0f18] __tracing_resize_ring_buffer+0x38/0xe0
  [c751bec0] [c00f1198] tracing_entries_write+0xb8/0x140
  [c751bef0] [c01644d4] vfs_write+0xd4/0x1e0
  [c751bf10] [c0164b98] SyS_write+0x58/0xc0
  [c751bf40] [c0011190] ret_from_syscall+0x0/0x3c
  --- Exception: c00 at 0xff225f4
      LR = 0xfec6554

The original commit tried to resolve two issues:
  a) Inefficient invoke of schedule_work_on()
  b) The function ring_buffer_resize() may be invoked by
     tracing_snapshot_alloc() before work queues is initialized

For issue (a) we have to still use the schedule_work_on() even we are
resizing the ring buffer for the current cpu in order to fix the
above call trace.

For issue (b) the ftrace is only functional after the tracer buffer
is allocated. Both the function tracer_alloc_buffers() and
init_workqueues() are initialized as early_initcall, but the
workqueue.o is listed before the trace/trace.o in the kernel/Makefile.
This can make sure that ring_buffer_resize() is never invoked before
the work queue init. So we should never run into a situation like this.

So it is safe to revert this commit.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>