rt: drop patch that breaks irqsafe operation
CQID: WIND00441708

One concern is that the problem the original patch was trying
to solve may re-appear.  It looks like it was a lockdep splat?
The patch has been present since 3.0-rt or earlier; it is not
in 3.8-rt.

In addition to the irqsafe breakage, this commit was the one
that introduced the horrible #ifdef within if mess.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>