move the NO_HZ to the standard.cfg
CQID: WIND00443000

The NO_HZ is now a mature feature and works pretty well on all archs.
So we should move it to a common cfg to avoid the duplication of
enabling it in every BSP. Since we only support this feature on
standard kernel, add it to standard.cfg and remove all the implicitly
or explicitly inclusion of it in other cfg files.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>