refresh usb-dwc3-add-keystone-usb-dwc3-support.patch
This refresh should be considered suspect, and requires developer
attention.

The commit 734b2fe93f4bcce29f7d971a106ea94e263e8676, "usb: dwc3: gadget:
don't request IRQs in atomic", in the stable tree makes changes to
dwc3_gadget_start() and dwc3_gadget_stop() which are relevant to this
patch.

I've transposed the failing hunks in this patch, but it's probably not
right.  It is difficult to say what the right thing is, especially
because this patch combines several commits.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>