ARM: OMAP3: Fix USB PHY Power-On Issue on Beagleboard in Kernel 3.10
1/1 [
Author: Sebastian Lenartowicz
Email: Sebastian.Lenartowicz@windriver.com
Subject: ARM: OMAP3: Fix USB PHY Power-On Issue on Beagleboard in Kernel 3.10
Date: Wed, 6 Nov 2013 13:48:56 -0500

Changes in ehci-omap altered how power-on had to be handled for
peripherals. This fix ensures that the power for the USB PHY on the
Beagleboard is on at runtime.

Tested on a Beagleboard xM Rev. C2.

Signed-off-by: Sebastian Lenartowicz <Sebastian.Lenartowicz@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>