intel-x86-64: restore KMACHINE AND KARCH specs to leaf scc nodes
The kernel-tools will pickup the file containing these definitions, so
it has to be the lead node that specifies them.

Additonally, include intel-x86-64.scc appears missing, or maybe that
was by design; in anycase the leaf nodes intel-x86-64-standard.scc and
intel-x86-64-preempt-rt.scc are restored here.

Put the branch standard at the start of the frag intel-x86-64.scc, to be
safe, so future includes that do have patches will follow branch.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>