intel-atom: drop gobinet module
CQID: WIND00443867

This is a non-mainline usb network driver that isn't properly coded
up to handle "# CONFIG_PM is not set" builds.  Since USB is by
definition not onboard, and since it was a module and not built in,
it clearly doesn't need to exist at all.  So simply drop it from
the builds.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>