net: iwlwifi: request only a threaded handler for interrupts
WIND: US18910

1/1 [
Author: Sebastian Andrzej Siewior
Email: bigeasy@linutronix.de
Subject: net: iwlwifi: request only a threaded handler for interrupts
Date: Wed, 27 Nov 2013 21:04:20 -0800

On RT the trans_pcie->irq_lock lock is converted into a sleeping lock
and can't be used in primary irq handler. The lock is used in mutliple
places which means turning it into a raw lock could increase the
latency of the system.
For now both handlers are moved into the thread.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Zhong Hongbo <hongbo.zhong@windriver.com>
]

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>