security: don't alloc the credential for the non-implemented function in cap_task_prctl
CQID: WIND00446184

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: security: don't alloc the credential for the non-implemented function in cap_task_prctl
Date: Wed, 18 Dec 2013 09:31:23 +0800

In function cap_task_prctl(), we would allocate a credential
unconditionally and then check if we support the requested function.
If not we would release this credential with abort_creds() by using
RCU method. But on some archs such as powerpc, the sys_prctl is heavily
used to get/set the floating point exception mode. So the unnecessary
allocating/releasing of credential not only introduce runtime overhead
but also do cause OOM due to the RCU implementation.  So we should
return immediately for the non-implemented function in the
cap_task_prctl().

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>