ACP34xx:Drop the codes to check the 'enabled' property in DTB for PCI driver
CQID: US29425

1/1 [
Author: Jiang Lu
Email: lu.jiang@windriver.com
Subject: ACP34xx:Drop the codes to check the 'enabled' property in DTB for PCI driver
Date: Fri, 17 Jan 2014 10:57:29 +0800

We have already changed to check the 'status' property instead of
'enabled' in the DTB for the PCI driver. But forgot to drop the
corresponding codes about the checking of 'enabled'. Just remove them.

Signed-off-by: Jiang Lu <lu.jiang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>