wireless: radiotap: fix parsing buffer overrun - linux - CVE-2013-7027
JIRAID: LIN5-13605

1/1 [
Author: Johannes Berg
Email: johannes.berg@intel.com
Subject: wireless: radiotap: fix parsing buffer overrun
Date: Fri, 10 Jan 2014 10:32:32 +0800

commit f5563318ff1bde15b10e736e97ffce13be08bc1a upstream

When parsing an invalid radiotap header, the parser can overrun
the buffer that is passed in because it doesn't correctly check
 1) the minimum radiotap header size
 2) the space for extended bitmaps

The first issue doesn't affect any in-kernel user as they all
check the minimum size before calling the radiotap function.
The second issue could potentially affect the kernel if an skb
is passed in that consists only of the radiotap header with a
lot of extended bitmaps that extend past the SKB. In that case
a read-only buffer overrun by at most 4 bytes is possible.

Fix this by adding the appropriate checks to the parser.

Cc: stable@vger.kernel.org
Reported-by: Evan Huus <eapache@gmail.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>