Security Advisory - linux - CVE-2013-4470
CQID: WIND00444028
JIRAID: LIN3-28899

1/2 [
Author: Hannes Frederic Sowa
Email: hannes@stressinduktion.org
Subject: ipv6: udp packets following an UFO enqueued packet need also be handled by UFO
Date: Mon, 11 Nov 2013 17:02:12 +0800

commit 2811ebac2521ceac84f2bdae402455baa6a7fb47 upstream

In the following scenario the socket is corked:
If the first UDP packet is larger then the mtu we try to append it to the
write queue via ip6_ufo_append_data. A following packet, which is smaller
than the mtu would be appended to the already queued up gso-skb via
plain ip6_append_data. This causes random memory corruptions.

In ip6_ufo_append_data we also have to be careful to not queue up the
same skb multiple times. So setup the gso frame only when no first skb
is available.

This also fixes a shortcoming where we add the current packet's length to
cork->length but return early because of a packet > mtu with dontfrag set
(instead of sutracting it again).

Found with trinity.

Cc: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Reported-by: Dmitry Vyukov <dvyukov@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Drop dontfrag part since we don't support IPV6_DONTFRAG feature.]
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

2/2 [
Author: Jiri Pirko
Email: jiri@resnulli.us
Subject: ip6_output: do skb ufo init for peeked non ufo skb as well
Date: Wed, 20 Nov 2013 18:02:15 +0800

commit c547dbf55d5f8cf615ccc0e7265e98db27d3fb8b upstream

Now, if user application does:
sendto len<mtu flag MSG_MORE
sendto len>mtu flag 0
The skb is not treated as fragmented one because it is not initialized
that way. So move the initialization to fix this.

introduced by:
commit e89e9cf539a28df7d0eb1d0a545368e9920b34ac "[IPv4/IPv6]: UFO Scatter-gather approach"

Signed-off-by: Jiri Pirko <jiri@resnulli.us>
Acked-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Adjust the codes to meet 2.6.27 kernel's need.]
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>