ipvs: Add boundary check on ioctl arguments - linux - CVE-2013-4588
CQID: WIND00445843
JIRAID: LIN3-15121

1/1 [
Author: Arjan van de Ven
Email: arjan@linux.intel.com
Subject: ipvs: Add boundary check on ioctl arguments
Date: Tue, 7 Jan 2014 13:47:34 +0800

commit 04bcef2a83f40c6db24222b27a52892cba39dffb upstream

The ipvs code has a nifty system for doing the size of ioctl command
copies; it defines an array with values into which it indexes the cmd
to find the right length.

Unfortunately, the ipvs code forgot to check if the cmd was in the
range that the array provides, allowing for an index outside of the
array, which then gives a "garbage" result into the length, which
then gets used for copying into a stack buffer.

Fix this by adding sanity checks on these as well as the copy size.

[ horms@verge.net.au: adjusted limit to IP_VS_SO_GET_MAX ]
Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Acked-by: Julian Anastasov <ja@ssi.bg>
Signed-off-by: Simon Horman <horms@verge.net.au>
Signed-off-by: Patrick McHardy <kaber@trash.net>
[Make the change on ipv4/ipvs/ip_vs_ctl.c instead of on
net/netfilter/ipvs/ip_vs_ctl.c.]
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>