intel-common: Add intel-common BSPs
Create the intel-core2-32 and intel-corei7-64 BSP descriptions. These
BSPs include all the core support for the other Intel BSPs in the
repository by including the corresponding BSP scc file.

This is an initial step to get the machines available and testing.
Further refactoring is expected to take place to reduce duplication and
ultimately obviate the need for many of the other BSP descriptions, at
least the -standard versions which should be adequately covered by these
generic versions.

Signed-off-by: Darren Hart <dvhart@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>