ipv6: fix headroom calculation in udp6_ufo_fragment - linux - CVE-2013-4563
CQID: WIND00445841
JIRAID: LIN6-4979

1/1 [
Author: Hannes Frederic Sowa
Email: hannes@stressinduktion.org
Subject: ipv6: fix headroom calculation in udp6_ufo_fragment
Date: Tue, 7 Jan 2014 14:15:11 +0800

commit 0e033e04c2678dbbe74a46b23fffb7bb918c288e upstream

Commit 1e2bd517c108816220f262d7954b697af03b5f9c ("udp6: Fix udp
fragmentation for tunnel traffic.") changed the calculation if
there is enough space to include a fragment header in the skb from a
skb->mac_header dervived one to skb_headroom. Because we already peeled
off the skb to transport_header this is wrong. Change this back to check
if we have enough room before the mac_header.

This fixes a panic Saran Neti reported. He used the tbf scheduler which
skb_gso_segments the skb. The offsets get negative and we panic in memcpy
because the skb was erroneously not expanded at the head.

Reported-by: Saran Neti <Saran.Neti@telus.com>
Cc: Pravin B Shelar <pshelar@nicira.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>