inet: fix addr_len/msg->msg_namelen assignment in recv_error and rxpmtu functions
JIRAID: LIN4-30958 LIN4-30966 LIN4-30969

1/2 [
Author: Hannes Frederic Sowa
Email: hannes@stressinduktion.org
Subject: inet: prevent leakage of uninitialized memory to user in recv syscalls
Date: Mon, 20 Jan 2014 15:24:24 +0800

commit bceaa90240b6019ed73b49965eac7d167610be69 upstream

Only update *addr_len when we actually fill in sockaddr, otherwise we
can return uninitialized memory from the stack to the caller in the
recvfrom, recvmmsg and recvmsg syscalls. Drop the the (addr_len == NULL)
checks because we only get called with a valid addr_len pointer either
from sock_common_recvmsg or inet_recvmsg.

If a blocking read waits on a socket which is concurrently shut down we
now return zero and set msg_msgnamelen to 0.

Reported-by: mpb <mpb.mail@gmail.com>
Suggested-by: Eric Dumazet <eric.dumazet@gmail.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
[XZ: Drop the unneeded ieee802154 dgram and ipv4 ping parts.]
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

2/2 [
Author: Hannes Frederic Sowa
Email: hannes@stressinduktion.org
Subject: inet: fix addr_len/msg->msg_namelen assignment in recv_error and rxpmtu functions
Date: Sun, 26 Jan 2014 10:58:14 +0800

commit 85fbaa75037d0b6b786ff18658ddf0b4014ce2a4 upstream

Commit bceaa90240b6019ed73b49965eac7d167610be69 ("inet: prevent leakage
of uninitialized memory to user in recv syscalls") conditionally updated
addr_len if the msg_name is written to. The recv_error and rxpmtu
functions relied on the recvmsg functions to set up addr_len before.

As this does not happen any more we have to pass addr_len to those
functions as well and set it to the size of the corresponding sockaddr
length.

This broke traceroute and such.

Fixes: bceaa90240b6 ("inet: prevent leakage of uninitialized memory to user in recv syscalls")
Reported-by: Brad Spengler <spender@grsecurity.net>
Reported-by: Tom Labanowski
Cc: mpb <mpb.mail@gmail.com>
Cc: David S. Miller <davem@davemloft.net>
Cc: Eric Dumazet <eric.dumazet@gmail.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
[XZ: Drop the changes for ipv4/ipv6 ping.c and ipv6_recv_rxpmtu().]
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>