xps: Improvements in TX queue selection
JIRAID: LIN4-7067

1/2 [
Author: Helmut Schaa
Email: helmut.schaa@googlemail.com
Subject: net: fix tx queue selection for bridged devices implementing select_queue
Date: Thu, 16 Jan 2014 15:01:02 +0800

commit deabc772f39405054a438d711f408d2d94d26d96 upstream.

When a net device is implementing the select_queue callback and is part of
a bridge, frames coming from the bridge already have a tx queue associated
to the socket (introduced in commit a4ee3ce3293dc931fab19beb472a8bde1295aebe,
"net: Use sk_tx_queue_mapping for connected sockets"). The call to
sk_tx_queue_get will then return the tx queue used by the bridge instead
of calling the select_queue callback.

In case of mac80211 this broke QoS which is implemented by using the
select_queue callback. Furthermore it introduced problems with rt2x00
because frames with the same TID and RA sometimes appeared on different
tx queues which the hw cannot handle correctly.

Fix this by always calling select_queue first if it is available and only
afterwards use the socket tx queue mapping.

Signed-off-by: Helmut Schaa <helmut.schaa@googlemail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Baogen Shang <baogen.shang@windriver.com>
]

2/2 [
Author: Tom Herbert
Email: therbert@google.com
Subject: xps: Improvements in TX queue selection
Date: Thu, 16 Jan 2014 15:17:12 +0800

commit 3853b5841c01a3f492fe137afaad9c209e5162c6 upstream.

In dev_pick_tx, don't do work in calculating queue
index or setting
the index in the sock unless the device has more than one queue.  This
allows the sock to be set only with a queue index of a multi-queue
device which is desirable if device are stacked like in a tunnel.

We also allow the mapping of a socket to queue to be changed.  To
maintain in order packet transmission a flag (ooo_okay) has been
added to the sk_buff structure.  If a transport layer sets this flag
on a packet, the transmit queue can be changed for the socket.
Presumably, the transport would set this if there was no possbility
of creating OOO packets (for instance, there are no packets in flight
for the socket).  This patch includes the modification in TCP output
for setting this flag.

Signed-off-by: Tom Herbert <therbert@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Baogen Shang <baogen.shang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>