gianfar/pm: fix hibernation failure
JIRAID: LIN6-6616

1/1 [
Author: Tiejun Chen
Email: tiejun.chen@windriver.com
Subject: gianfar/pm: fix hibernation failure
Date: Thu, 27 Feb 2014 10:45:23 +0800

This is an issue caused by the mismatch between gfar_suspend()
and gfar_restore().

Before entering the hibernation the gfar_suspend() will release
the skb and irq resources used by gfar if the wol is not enabled.
When leaving hibernation, the gfar_restore() is invoked instead
of gfar_resume() and it assumes that the skb buffer is still
available. This is definitely not true and then causes the kernel
panic.

So change the gfar_suspend() and gfar_release() not to release
the skb and irq resources no matter wol is enabled or not.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>