net: fix info leak in compat dev_ifconf() - linux - CVE-2012-6539
CQID: WIND00413572
JIRAID: LIN4-23319

1/1 [
Author: Mathias Krause
Email: minipli@googlemail.com
Subject: net: fix info leak in compat dev_ifconf()
Date: Fri, 7 Mar 2014 16:30:00 +0800

commit 43da5f2e0d0c69ded3d51907d9552310a6b545e8 upstream

The implementation of dev_ifconf() for the compat ioctl interface uses
an intermediate ifc structure allocated in userland for the duration of
the syscall. Though, it fails to initialize the padding bytes inserted
for alignment and that for leaks four bytes of kernel stack. Add an
explicit memset(0) before filling the structure to avoid the info leak.

Signed-off-by: Mathias Krause <minipli@googlemail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>