drivers/cdrom/cdrom.c: use kzalloc() for failing hardware - linux - CVE-2013-2164
CQID: WIND00426609
JIRAID: LIN4-25726

1/1 [
Author: yanjun.zhu
Email: yanjun.zhu@windriver.com
Subject: drivers/cdrom/cdrom.c: use kzalloc() for failing hardware
Date: Wed, 12 Mar 2014 14:05:11 +0800

commit 050e4b8fb7cdd7096c987a9cd556029c622c7fe2 upstream

In drivers/cdrom/cdrom.c mmc_ioctl_cdrom_read_data() allocates a memory
area with kmalloc in line 2885.

2885         cgc->buffer = kmalloc(blocksize, GFP_KERNEL);
2886         if (cgc->buffer == NULL)
2887                 return -ENOMEM;

In line 2908 we can find the copy_to_user function:

2908         if (!ret && copy_to_user(arg, cgc->buffer, blocksize))

The cgc->buffer is never cleaned and initialized before this function.  If
ret = 0 with the previous basic block, it's possible to display some
memory bytes in kernel space from userspace.

When we read a block from the disk it normally fills the ->buffer but if
the drive is malfunctioning there is a chance that it would only be
partially filled.  The result is an leak information to userspace.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: yanjun.zhu <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>