s390: fix kernel crash due to linkage stack instructions - linux - CVE-2014-2039
JIRAID: LIN5-18193

1/1 [
Author: Martin Schwidefsky
Email: schwidefsky@de.ibm.com
Subject: s390: fix kernel crash due to linkage stack instructions
Date: Mon, 3 Feb 2014 17:37:15 +0100

commit 8d7f6690cedb83456edd41c9bd583783f0703bf0 upstream

The kernel currently crashes with a low-address-protection exception
if a user space process executes an instruction that tries to use the
linkage stack. Set the base-ASTE origin and the subspace-ASTE origin
of the dispatchable-unit-control-table to point to a dummy ASTE.
Set up control register 15 to point to an empty linkage stack with no
room left.

A user space process with a linkage stack instruction will still crash
but with a different exception which is correctly translated to a
segmentation fault instead of a kernel oops.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Yong Zhang <yong.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>