bnx2x: remove false warning regarding interrupt number
JIRAID: US36762

1/1 [
Author: Ariel Elior
Email: ariele@broadcom.com
Subject: bnx2x: remove false warning regarding interrupt number
Date: Thu, 20 Sep 2012 05:26:41 +0000

commit 185d4c8bf579322e1c2835d70729bc30f6f80f55 upstream

Since version 7.4 the FW configures in the pci config space the max
number of interrupts available to the physical function, instead of
the exact number to use.
This causes a false warning in driver when comparing the number of
configured interrupts to the number about to be used.

Signed-off-by: Ariel Elior <ariele@broadcom.com>
Signed-off-by: Eilon Greenstein <eilong@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>