arm/dts: limit the max frame size up to 3800
JIRAID: LIN5-18533

1/1 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: arm/dts: limit the max frame size up to 3800
Date: Tue, 29 Apr 2014 14:41:28 +0800

Commit 0e899ace0777818e2fd8fd99d2c4449948ad375f
and    7850c5f50246d1e21f439bd6c8da8aec722ad563
>From http://rocketboards.org/gitweb/?p=linux-socfpga.git

Commit 0e899ace0777818e2fd8fd99d2c4449948ad375f create this
limitation as 4062 as below:

The valid range for the C5 SOC is 2000 through 4062, inclusive.

Commit 7850c5f50246d1e21f439bd6c8da8aec722ad563 shrink it to 3800:

The Synopsys EMAC will checksum jumbo transmit frames only up to a
certain size correctly, since the controller will start to egress
an outbound packet when it starts reaching the maximum size of the
transmit fifo. In the Altera configuration case, the PBL, or
programmed burst length is set to 64, the transmit fifo size is 4K,
and the bus datawidth is 32-bits. The equation to determine the
correct maximum MTU is FIFOSIZE-((PBL+3)*(BUSWIDTH/8)) ==
4096-(67*4) = 3828. Then we need to account for possible VLAN tags,
inner and outer, and then transform than number into an Ethernet
MTU, which gives us 3828-8-14-4 (8 bytes for two VLANs, 14 bytes
for the standard Ethernet header size, and another 4 bytes for
CRC). That gives us 3802. So we'll make it a maximum jumbo frame
size of 3800.

Signed-off-by: Vince Bridgers <vbridger@altera.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>