System hangs by BUG_ON() in the brd driver caused by a small race in brd_lookup_page()
JIRAID: LIN4-31287

1/1 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: System hangs by BUG_ON() in the brd driver caused by a small race in brd_lookup_page()
Date: Tue, 29 Apr 2014 11:25:50 +0800

commit dfd20b2b174d3a9b258ea3b7a35ead33576587b1 upstream

The index on the page must be set before it is inserted in the
radix tree.  Otherwise there is a small race which can occur
during lookup where the page can be found with the incorrect
index.  This will trigger the BUG_ON() in brd_lookup_page().

Signed-off-by: Andrew Morton <akpm@linux-foundation.org>,
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>